requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis for caching
redis==5.0.1

# HTTP requests
httpx==0.25.2
requests==2.31.0

# AI and NLP
openai==1.3.7

# Environment variables
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# JSON handling
pydantic==2.5.0

# CORS
fastapi-cors==0.0.6

# Background tasks
celery==5.3.4

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0 